batctl: export generic bisect functionalities
Some of the code in the bisect_iv module can be reused to
bisect other protocols (e.g. B.A.T.M.A.N. V).

Therefore such generic routines can be exported in a common
module which can be reused by new bisectors.

Functions, structure and variables have been renamed in
order to have a better-looking code and to respect the
namespace.

In this patch there is no behavioural change or code
generalisation.

Signed-off-by: Antonio Quartulli <antonio@open-mesh.com>